backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-messages==0.6.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==3.3
mysqlclient==2.0.3
numpy==1.21.3
oauthlib==3.1.1
pandas==1.3.4
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
Pillow==8.4.0